backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
beautifulsoup4==4.13.4
pydantic-settings==2.10.1
eval-type-backport==0.2.2

pytest==8.4.1
pytest-asyncio==1.1.0
asgi-lifespan==2.1.0
